[SW-56] refactor(mission): MissionType 구조 최적화
