+++ src/main/java/com/github/kmu_shell_we/domain/mission/constant/MissionType.java
@@ -1,7 +1,9 @@
-package com.github.kmu_shell_we.domain.mission.dto.response;
+package com.github.kmu_shell_we.domain.mission.constant;
 
 public enum MissionType {
+
     DAILY,
     WEEKLY,
     SPECIAL
 }
+
