+++ src/main/java/com/github/kmu_shell_we/domain/mission/entity/Mission.java
@@ -1,12 +1,12 @@
 package com.github.kmu_shell_we.domain.mission.entity;
 
+import com.github.kmu_shell_we.domain.mission.constant.MissionType;
 import com.github.kmu_shell_we.global.infra.mysql.BaseSchema;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
 import lombok.*;
 import lombok.experimental.SuperBuilder;
 
-import java.nio.charset.Charset;
 
 @Entity
 @Getter
@@ -28,12 +28,5 @@ public class Mission extends BaseSchema {
     @Column(nullable = false)
     @NotNull
     @Enumerated(EnumType.STRING)
-    Mission.MissionType type;
-
-    public enum MissionType {
-
-        DAILY,
-        WEEKLY,
-        SPECIAL,
-    }
+    MissionType type;
 }
