+++ src/main/java/com/github/kmu_shell_we/domain/mission/service/MissionService.java
@@ -3,7 +3,6 @@ package com.github.kmu_shell_we.domain.mission.service;
 import com.github.kmu_shell_we.domain.mission.dto.request.UpsertMissionRequest;
 import com.github.kmu_shell_we.domain.mission.dto.response.MissionListResponse;
 import com.github.kmu_shell_we.domain.mission.dto.response.MissionResponse;
-import com.github.kmu_shell_we.domain.mission.dto.response.MissionType;
 import com.github.kmu_shell_we.domain.mission.entity.Mission;
 import com.github.kmu_shell_we.domain.mission.exception.MissionExceptionCode;
 import com.github.kmu_shell_we.domain.mission.repository.MissionRepository;
@@ -39,7 +38,7 @@ public class MissionService {
                 Mission.builder()
                         .name(request.getName())
                         .reward(request.getReward())
-                        .type(Mission.MissionType.valueOf(request.getType().name()))
+                        .type(request.getType())
                         .build()
         );
 
@@ -55,7 +54,7 @@ public class MissionService {
                         .toBuilder()
                         .name(request.getName())
                         .reward(request.getReward())
-                        .type(Mission.MissionType.valueOf(request.getType().name()))
+                        .type(request.getType())
                         .build()
         );
